Refactor InfinispanCacheConfiguration/RemoteCacheConfiguration into wildfly-clustering-ee-* modules so they can be reused across web/ejb modules.
