+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/sso/hotrod/HotRodSSOManagementConfiguration.java
@@ -22,7 +22,7 @@
 
 package org.wildfly.extension.clustering.web.sso.hotrod;
 
-import org.wildfly.clustering.web.hotrod.RemoteCacheConfiguration;
+import org.wildfly.clustering.ee.hotrod.RemoteCacheConfiguration;
 
 /**
  * @author Paul Ferraro
